requirements.txt
Django>=4.0.0,<4.1.0
yt-dlp
djangorestframework
python-dotenv
requests
ipython
gunicorn
dj_database_url
psycopg2
markdown
SpeechRecognition
django-storages
boto3
django-rename-app
